Add opencpu error handling and additional testing

+++ metabulo/views.py
@@ -1,9 +1,9 @@
 from io import BytesIO
 
-from flask import Blueprint, current_app, jsonify, request, Response, send_file
+from flask import Blueprint, current_app, jsonify, request, send_file
 from sklearn import preprocessing
 
-from metabulo.models import CSVFile, CSVFileSchema, db
+from metabulo.models import CSVColumn, CSVFile, CSVFileSchema, db
 from metabulo.opencpu import process_table
 
 
@@ -41,6 +41,7 @@ def upload_csv_file():
 
 @csv_bp.route('/csv', methods=['POST'])
 def create_csv_file():
+    csv_file = None
     try:
         csv_file = csv_file_schema.load(request.json)
         db.session.add(csv_file)
@@ -48,7 +49,7 @@ def create_csv_file():
 
         return jsonify(csv_file_schema.dump(csv_file)), 201
     except Exception:
-        if csv_file.uri.is_file():
+        if csv_file and csv_file.uri.is_file():
             csv_file.uri.unlink()
         db.session.rollback()
         raise
@@ -70,7 +71,8 @@ def download_csv_file(csv_id):
 @csv_bp.route('/csv/<uuid:csv_id>', methods=['DELETE'])
 def delete_csv_file(csv_id):
     csv_file = CSVFile.query.get_or_404(csv_id)
-    csv_file.delete()
+    CSVColumn.query.filter_by(file_id=csv_id).delete()
+    db.session.delete(csv_file)
     db.session.commit()
 
     try:
@@ -108,15 +110,11 @@ def fill_missing_values(csv_id, column):
 def echo_csv_file(csv_id):
     csv_file = CSVFile.query.get_or_404(csv_id)
     table = csv_file.table
-    api_root = current_app.config['OPENCPU_API_ROOT']
-    echo = process_table(api_root + '/metabulo/R/echo', table)
-    return Response(echo.to_csv(), mimetype='text/csv')
+    return process_table('/metabulo/R/echo', table)
 
 
 @csv_bp.route('/csv/<uuid:csv_id>/demo', methods=['GET'])
 def processing_demo(csv_id):
     csv_file = CSVFile.query.get_or_404(csv_id)
     table = csv_file.table
-    api_root = current_app.config['OPENCPU_API_ROOT']
-    echo = process_table(api_root + '/metabulo/R/demo', table)
-    return Response(echo.to_csv(), mimetype='text/csv')
+    return process_table('/metabulo/R/demo', table)
